Resources: Add Notes directory and initial notes
These are my notes from the first night of Capstone, as well
as the first night of Project presentations.  The resources to
compile the LaTeX are all located in the added Notes directory,
and a copy of the compiled notes are in the main Capstone-Resources
directory as well.  You can view them either in the Notes directory
as main.pdf, or in the main directory as Notes.pdf

Note: I use pdflatex to compile all of my .tex files locally, if
      you would like to compile them yourself you just run the
      command 'pdflatex main.tex'

Signed-off-by: Alice Michael <malice@pdx.edu>