Resources: update README with merging commands
This updates the README to have the merging command to pull
master into your branch.  As I just updated the file name to
be different, it will be a clean merge.  This is for people
to pull the notes into their local branch if wanted, and
is to learn more about tree management and merging.

Signed-off-by: Alice Michael <malice@pdx.edu>